Consolidate orbit-store into orbit package.
