--- src/orbit-store.js
@@ -1,4 +1,0 @@
-import Store from 'orbit-store/store';
-import Cache from 'orbit-store/cache';
-
-export { Store, Cache };
