+++ test/tests/orbit/unit/cache-test.js
@@ -16,7 +16,7 @@ import {
   // replaceHasMany,
   replaceHasOne
 } from 'orbit/transform/operators';
-import Cache from 'orbit-store/cache';
+import Cache from 'orbit/cache';
 
 let schema, keyMap;
